+++ tests/leb-freeze/test_calibration.py
@@ -1,0 +1,5 @@
+import leb.freeze.calibration
+
+
+def test_calibration():
+    raise NotImplementedError