+++ CHANGELOG.md
@@ -5,17 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [uManager-Unreleased]
+## [leb-freeze-Unreleased]
 
 ### Added
 
-- Added a `calibrate.bsh` script to run the Fourier ptychography calibration routine.
+- Added leb-freeze Python package to analyze Fourier ptychography data.
 
 ## [LED-Matrix-Controller-0.0.0]
 
 ### Added
 
 - Responses from the Arduino to the PC are now terminated with the linefeed `\n` character.
+
+## [uManager-Unreleased]
+
+### Added
+
+- Added a `calibrate.bsh` script to run the Fourier ptychography calibration routine.
 
 
 [uManager-Unreleased]: https://github.com/leb-epfl/mr-freeze/